Refactor ExpandableListViewEventAdapters
