+++ app/src/main/java/com/github/dedis/student20_pop/utility/ui/EventAdapter/OrganizerExpandableListViewEventAdapter.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.utility.ui.organizer;
+package com.github.dedis.student20_pop.utility.ui.EventAdapter;
 
 import android.app.AlertDialog;
 import android.content.Context;
@@ -6,26 +6,23 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.BaseExpandableListAdapter;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
 import com.github.dedis.student20_pop.R;
 import com.github.dedis.student20_pop.model.Event;
+import com.github.dedis.student20_pop.utility.ui.organizer.OnEventTypeSelectedListener;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
+
+import static com.github.dedis.student20_pop.model.Event.EventCategory;
+import static com.github.dedis.student20_pop.model.Event.EventCategory.FUTURE;
+import static com.github.dedis.student20_pop.model.Event.EventType.*;
 
 /**
  * Adapter to show events of an Organizer
  */
-public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListAdapter {
-    private final Context context;
-    private final List<EventCategory> categories;
-    private final HashMap<EventCategory, List<Event>> eventsMap;
+public class OrganizerExpandableListViewEventAdapter extends ExpandableListViewEventAdapter {
     private final OnEventTypeSelectedListener onEventTypeSelectedListener;
 
     /**
@@ -36,82 +33,10 @@ public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListA
      * @param events  the list of events of the lao
      */
     public OrganizerExpandableListViewEventAdapter(Context context, List<Event> events, OnEventTypeSelectedListener onEventTypeSelectedListener) {
-        this.context = context;
-        this.eventsMap = new HashMap<>();
-        this.categories = new ArrayList<>();
-        this.categories.add(EventCategory.PAST);
-        this.categories.add(EventCategory.PRESENT);
-        this.categories.add(EventCategory.FUTURE);
-        this.eventsMap.put(EventCategory.PAST, new ArrayList<>());
-        this.eventsMap.put(EventCategory.PRESENT, new ArrayList<>());
-        this.eventsMap.put(EventCategory.FUTURE, new ArrayList<>());
+        super(context, events);
         this.onEventTypeSelectedListener = onEventTypeSelectedListener;
-
-        putEventsInMap(events, this.eventsMap);
-        orderEventsInMap(this.eventsMap);
     }
 
-    /**
-     * @return the amount of categories
-     */
-    @Override
-    public int getGroupCount() {
-        return this.eventsMap.size();
-    }
-
-    /**
-     * @param groupPosition
-     * @return the amount of events in a given group
-     */
-    @Override
-    public int getChildrenCount(int groupPosition) {
-        return this.eventsMap.get(this.categories.get(groupPosition)).size();
-    }
-
-    /**
-     * @param groupPosition
-     * @return the category of a given position
-     */
-    @Override
-    public Object getGroup(int groupPosition) {
-        if (groupPosition >= getGroupCount()) {
-            return null;
-        }
-        return this.categories.get(groupPosition);
-    }
-
-    /**
-     * @param groupPosition
-     * @param childPosition
-     * @return the event for a given position in a given category
-     */
-    @Override
-    public Object getChild(int groupPosition, int childPosition) {
-
-        if (groupPosition >= getGroupCount()) {
-            return null;
-        }
-        if (childPosition >= getChildrenCount(groupPosition)) {
-            return null;
-        }
-
-        return this.eventsMap.get(this.categories.get(groupPosition)).get(childPosition);
-    }
-
-    @Override
-    public long getGroupId(int groupPosition) {
-        return groupPosition;
-    }
-
-    @Override
-    public long getChildId(int groupPosition, int childPosition) {
-        return childPosition;
-    }
-
-    @Override
-    public boolean hasStableIds() {
-        return false;
-    }
 
     /**
      * @param groupPosition
@@ -144,7 +69,7 @@ public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListA
         eventTextView.setText(eventCategory);
 
         ImageButton addEvent = convertView.findViewById(R.id.add_future_event_button);
-        addEvent.setVisibility((getGroup(groupPosition) == EventCategory.FUTURE) ? View.VISIBLE : View.GONE);
+        addEvent.setVisibility((getGroup(groupPosition) == FUTURE) ? View.VISIBLE : View.GONE);
         addEvent.setFocusable(View.NOT_FOCUSABLE);
         addEvent.setOnClickListener(v -> {
             AlertDialog.Builder builderSingle = new AlertDialog.Builder(context);
@@ -152,9 +77,9 @@ public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListA
 
             final ArrayAdapter<String> arrayAdapter = new ArrayAdapter<>(context, android.R.layout.select_dialog_singlechoice);
 
-            arrayAdapter.insert(context.getString(R.string.meeting_event), Event.EventType.MEETING.ordinal());
-            arrayAdapter.insert(context.getString(R.string.roll_call_event), Event.EventType.ROLL_CALL.ordinal());
-            arrayAdapter.insert(context.getString(R.string.poll_event), Event.EventType.POLL.ordinal());
+            arrayAdapter.insert(context.getString(R.string.meeting_event), MEETING.ordinal());
+            arrayAdapter.insert(context.getString(R.string.roll_call_event), ROLL_CALL.ordinal());
+            arrayAdapter.insert(context.getString(R.string.poll_event), POLL.ordinal());
 
             builderSingle.setNegativeButton(context.getString(R.string.button_cancel), (dialog, which) -> dialog.dismiss());
             builderSingle.setAdapter(arrayAdapter, (dialog, which) -> {
@@ -177,7 +102,8 @@ public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListA
      */
     @Override
     public View getChildView(int groupPosition, int childPosition, boolean isLastChild, View convertView, ViewGroup parent) {
-
+        //TODO : For the moment, events are displayed the same if user is attendee or organizer,
+        // in the future it could be nice to have a pencil icon to allow organizer to modify an event
         Event event = ((Event) getChild(groupPosition, childPosition));
         String eventTitle = (event.getName() + " : " + event.getType());
 
@@ -217,63 +143,5 @@ public class OrganizerExpandableListViewEventAdapter extends BaseExpandableListA
         TextView descriptionTextView = convertView.findViewById(R.id.event_description);
         descriptionTextView.setText(eventDescription);
         return convertView;
-    }
-
-    @Override
-    public boolean isChildSelectable(int groupPosition, int childPosition) {
-        return true;
-    }
-
-    /**
-     * A helper method that places the events in the correct key-value pair
-     * according to their times
-     *
-     * @param events
-     * @param eventsMap
-     */
-    private void putEventsInMap(List<Event> events, HashMap<EventCategory, List<Event>> eventsMap) {
-        //For now, the event are put in the different categories according to their time attribute
-        //Later, according to the start/end-time
-        for (Event event : events) {
-            //for now (testing purposes)
-            //later: event.getEndTime() < now
-            if (event.getTime() < 50) {
-                eventsMap.get(EventCategory.PAST).add(event);
-            }
-            //later: event.getStartTime()<now && event.getEndTime() > now
-            else if (event.getTime() < 1000) {
-                eventsMap.get(EventCategory.PRESENT).add(event);
-            } else { //if e.getStartTime() > now
-                eventsMap.get(EventCategory.FUTURE).add(event);
-            }
-        }
-    }
-
-    /**
-     * A helper method that orders the events according to their times
-     *
-     * @param eventsMap
-     */
-    private void orderEventsInMap(HashMap<EventCategory, List<Event>> eventsMap) {
-
-        for (EventCategory category : categories) {
-            Collections.sort(eventsMap.get(category), new EventComparator());
-        }
-        //2 possibilities: B strictly after A or B nested within A
-    }
-
-    /**
-     * Enum class for each event category
-     */
-    private enum EventCategory {
-        PAST, PRESENT, FUTURE
-    }
-
-    private static class EventComparator implements Comparator<Event> {
-        //later: compare start times
-        @Override
-        public int compare(Event event1, Event event2) {
-            return Long.compare(event1.getTime(), event2.getTime());
-        }
     }
 }
